added draw view and activity

+++ app/src/main/java/hu/bme/aut/stringartdesigner/model/Position.kt
@@ -1,6 +1,6 @@
 package hu.bme.aut.stringartdesigner.model
 
 data class Position (
-    var x: Int,
-    var y: Int
+    var x: Float,
+    var y: Float
 )
